# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Salles API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /auth/register:
    post:
      summary: Inscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, pseudo]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                pseudo: { type: string }
      responses:
        '201':
          description: Created
  /auth/login:
    post:
      summary: Connexion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
  /lieux:
    get:
      summary: Liste des lieux
      parameters:
        - { name: type, in: query, schema: { type: string, enum: [blacklist, favori] } }
        - { name: ville, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '200':
          description: OK
    post:
      summary: Créer un lieu
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LieuInput'
      responses:
        '201':
          description: Created
  /lieux/{id}:
    get:
      summary: Détail lieu
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: OK
    put:
      summary: Modifier lieu
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: OK
    delete:
      summary: Supprimer lieu
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '204':
          description: No Content
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LieuInput:
      type: object
      required: [nom, adresse]
      properties:
        nom: { type: string }
        adresse: { type: string }
        type: { type: string, enum: [blacklist, favori] }
        commentaire: { type: string }
        ville: { type: string }
        latitude: { type: number }
        longitude: { type: number }